Move splitter into editor directory
