+++ src/components/editor/editor.tsx
@@ -5,7 +5,7 @@ import useMedia from 'use-media'
 import { ApplicationState } from '../../redux'
 import { setEditorModeConfig } from '../../redux/editor/methods'
 import { DocumentTitle } from '../common/document-title/document-title'
-import { Splitter } from '../common/splitter/splitter'
+import { Splitter } from './splitter/splitter'
 import { MotdBanner } from '../common/motd-banner/motd-banner'
 import { DocumentBar } from './document-bar/document-bar'
 import { EditorWindow } from './editor-window/editor-window'
