+++ src/components/editor/splitter/split-divider/split-divider.scss
