+++ src/components/editor/splitter/splitter.scss
